Module container been added, attributes representation and more...
CPyModuleContainer, provides a mean to register and to own created modules. the update provides the following capabilities:
1) Register new modules.
2) Retrive existing modules.

CPythonFunction acts as a functor,
where the functor it self is the wrapper to call the stored user logic.

CPyException will verify if a python exception was already raised. in order not to overwrite it down.